requirements.txt
gunicorn==20.0.4
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
torchvision==0.7.0
bert-extractive-summarizer==0.5.0
Flask==1.1.1
Flask-Material==0.1.1
Jinja2==2.11.1
nltk==3.4.5
numpy==1.18.1
numpydoc==0.9.2
requests==2.22.0
rope==0.16.0
rouge==1.0.0
spacy==2.1.3
sumy==0.8.1
tokenizers==0.8.1rc1
transformers==2.2.2
newspaper3k==0.2.8
validator-collection==1.4.2